Moved settings class to settings.py file.
